convert dexterity to int

+++ include/Spell/FireSpell.h
@@ -12,7 +12,7 @@ class FireSpell : public Spell {
     public:
         FireSpell(string name) : Spell(name, FS_PRICE, FS_MIN_LEVEL, FS_MIN_DAMAGE, FS_MAX_DAMAGE, FS_MAGIC_POWER_REQUIRED) {}
         // Εκτός από τη ζημιά που προκαλεί, μειώνει και το ποσό άμυνας του αντιπάλου για κάποιους γύρους
-        void use(Monster* monster,  float dexterity);
+        void use(Monster* monster,  int dexterity);
 
         ~FireSpell() {}
