+++ include/Spell/LightingSpell.h
@@ -13,7 +13,7 @@ class LightningSpell : public Spell {
         LightningSpell(string name) : Spell(name, LS_PRICE, LS_MIN_LEVEL, LS_MIN_DAMAGE, LS_MAX_DAMAGE, LS_MAGIC_POWER_REQUIRED) {}
         /* Εκτός από τη ζημιά που προκαλεί,
             μειώνει και την πιθανότητα να αποϕύγει μια επίθεση ο αντίπαλος για κάποιους γύρους */
-        void use(Monster* monster,  float dexterity);
+        void use(Monster* monster,  int dexterity);
 
         ~LightningSpell(){};
